feat(achievements): Extract AchievementCard component from page
