+++ src/app/achievements/page.tsx
@@ -1,10 +1,6 @@
-import { AchievementCardBack } from '@/components/layout/achievements/achievement-card-back'
-import { AchievementCardFront } from '@/components/layout/achievements/achievement-card-front'
-import { FlippingCard } from '@/components/ui/flipping-card'
+import { AchievementCard } from '@/components/layout/achievements/achievement-card'
 import { SectionLabel } from '@/components/ui/section-label'
-import unknownAchievementImage from '@/images/achievements/unknown.webp'
-import { ACHIEVEMENTS, ACHIEVEMENTS_COOKIE_NAME, parseUnlockedCookie } from '@/lib/achievements'
-import { format, isValid as isValidDate, parseISO } from 'date-fns'
+import { ACHIEVEMENTS, ACHIEVEMENTS_COOKIE_NAME, parseUnlockedCookie, type AchievementId } from '@/lib/achievements'
 import { cookies } from 'next/headers'
 
 export default async function AchievementsPage() {
@@ -12,7 +8,7 @@ export default async function AchievementsPage() {
   const cookieStore = await cookies()
   const cookieValue = cookieStore.get(ACHIEVEMENTS_COOKIE_NAME)?.value
   const unlocked = parseUnlockedCookie(cookieValue)
-  const entries = Object.entries(ACHIEVEMENTS)
+  const entries: Array<[AchievementId, (typeof ACHIEVEMENTS)[AchievementId]]> = Object.entries(ACHIEVEMENTS)
 
   return (
     <div className="px-10 py-16 md:px-20 lg:px-40">
@@ -23,50 +19,11 @@ export default async function AchievementsPage() {
         </div>
 
         <ul className="grid grid-cols-1 gap-6 sm:grid-cols-2 lg:grid-cols-3">
-          {entries.map(([id, def]) => {
-            const unlockedAt = unlocked[id]
-            const isUnlocked = Boolean(unlockedAt)
-            const title = isUnlocked ? def.title : 'Hidden achievement'
-            const description = isUnlocked ? def.cardDescription : def.unlockHint
-            let footer = 'Keep exploring the site!'
-            if (isUnlocked) {
-              if (unlockedAt) {
-                // Safely parse and format. Prefer ISO parsing; fall back to Date constructor.
-                const isoDate = parseISO(unlockedAt)
-                if (isValidDate(isoDate)) {
-                  footer = `Unlocked: ${format(isoDate, 'PP p')}`
-                } else {
-                  const fallbackDate = new Date(unlockedAt)
-                  footer = Number.isNaN(fallbackDate.getTime())
-                    ? 'Unlocked'
-                    : `Unlocked: ${format(fallbackDate, 'PP p')}`
-                }
-              } else {
-                footer = 'Unlocked'
-              }
-            }
-            const imageSrc = isUnlocked ? def.imageSrc : unknownAchievementImage
-            const imageAlt = isUnlocked ? def.imageAlt : 'Unknown achievement'
-            const ariaLabel = isUnlocked ? `Achievement: ${def.title}` : 'Hidden achievement'
-
-            return (
-              <li key={id} className="list-none">
-                <FlippingCard
-                  front={<AchievementCardFront />}
-                  back={<AchievementCardBack title={title} description={description} footer={footer} />}
-                  backgroundImageSrc={imageSrc}
-                  backgroundImageAlt={imageAlt}
-                  backgroundSizes="(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw"
-                  ariaLabel={ariaLabel}
-                  className="rounded-xl"
-                  flipLabelFrontDesktop="View details"
-                  flipLabelFrontMobile="Flip"
-                  flipLabelBackDesktop="Go back"
-                  flipLabelBackMobile="Flip"
-                />
-              </li>
-            )
-          })}
+          {entries.map(([id]) => (
+            <li key={id} className="list-none">
+              <AchievementCard id={id} initialUnlockedAt={unlocked[id]} />
+            </li>
+          ))}
         </ul>
       </div>
     </div>
